IterativeAlgorithms.jl

AlgorithmsInterface.Algorithm — Type

Algorithm

An abstract type to represent an algorithm.

A concrete algorithm contains all static parameters that characterise the algorithms. Together with a Problem an Algorithm subtype should be able to initialize or reset a State.

Example

For a gradient descent algorithm the algorithm would specify which step size selection to use.

AlgorithmsInterface.Problem — Type

Problem

An abstract type to represent a problem to be solved with all its static properties, that do not change during an algorithm run.

Example

For a gradient descent algorithm the problem consists of

a cost function $f: C → ℝ$

a gradient function $\operatorname{grad}f$

The problem then could that these are given in four different forms

a function c = cost(x) and a gradient d = gradient(x)

a function c = cost(x) and an in-place gradient gradient!(d,x)

a combined cost-grad function (c,d) = costgrad(x)

a combined cost-grad function (c, d) = costgrad!(d, x) that computes the gradient in-place.

AlgorithmsInterface.State — Type

State

An abstract type to represent the state an iterative algorithm is in.

The state consists of any information that describes the current step the algorithm is in and keeps all information needed from one step to the next.

Usually this should include

iteration – the current iteration step $k$ that is is currently performed or was last performed

stopping_criterion – a StoppingCriterion that indicates whether the algorithm will stop after this iteration or has stopped.

iterate the current iterate $x^{(k)}$`.

These variable names given in this list are the defaults for which the accessors are implemented, such that if your concrete MyState <: State follows this convention, you do not have to reimplement their accessors.

AlgorithmsInterface.step! — Function

step!(p::Problem, a::Algorithm, s::State)

Perform the current step of an Algorithm a solving Problem p starting from State s.

AlgorithmsInterface.get_iteration — Function

get_iteration(s::State)

Return the current iteration a state either is currently performing or was last performed

The default assumes that the current iteration is stored in s.iteration.