Main Interface types

AlgorithmsInterface.initialize_state! — Method

s = initialize_state(p::Problem, a::Algorithm; kwargs...)
initialize_state!(p::Problem, a::Algorithm, s::State; kwargs...)

Initialize a State s base on a Problem p and an Algorithm. The kwargs... should allow to initialize for example the initial point. This can be done in-place of s, then only values that did change have to be provided.

AlgorithmsInterface.initialize_state — Method

s = initialize_state(p::Problem, a::Algorithm; kwargs...)
initialize_state!(p::Problem, a::Algorithm, s::State; kwargs...)

Initialize a State s base on a Problem p and an Algorithm. The kwargs... should allow to initialize for example the initial point. This can be done in-place of s, then only values that did change have to be provided.

AlgorithmsInterface.is_finished — Method

is_finished(p::Problem, a::Algorithm, s::State)

AlgorithmsInterface.solve! — Method

solve!(p::Problem, a::Algorithm, s::State; kwargs...)

Solve the Problem p using the Algorithm a working on the State.

All keyword arguments are passed to the initialize_state! function.

AlgorithmsInterface.solve — Method

solve(p::Problem, a::Algorithm; kwargs...)

Solve the Problem p using the Algorithm a. The keyword arguments kwargs... have to provide enough details such that the corresponding state initialisation initialize_state(p,a) returns a state.

By default this method continues to call solve!.

AlgorithmsInterface.step! — Method

step!(p::Problem, a::Algorithm, s::State)

Perform the current step of an Algorithm a solving Problem p starting from State s.

Algorithm

AlgorithmsInterface.Algorithm — Type

Algorithm

An abstract type to represent an algorithm.

A concrete algorithm contains all static parameters that characterise the algorithms. Together with a Problem an Algorithm subtype should be able to initialize or reset a State.

Example

For a gradient descent algorithm the algorithm would specify which step size selection to use.

Problem

AlgorithmsInterface.Problem — Type

Problem

An abstract type to represent a problem to be solved with all its static properties, that do not change during an algorithm run.

Example

For a gradient descent algorithm the problem consists of

a cost function $f: C → ℝ$

a gradient function $\operatorname{grad}f$

The problem then could that these are given in four different forms

a function c = cost(x) and a gradient d = gradient(x)

a function c = cost(x) and an in-place gradient gradient!(d,x)

a combined cost-grad function (c,d) = costgrad(x)

a combined cost-grad function (c, d) = costgrad!(d, x) that computes the gradient in-place.

State

AlgorithmsInterface.State — Type

State

An abstract type to represent the state an iterative algorithm is in.

The state consists of any information that describes the current step the algorithm is in and keeps all information needed from one step to the next.

Usually this should include

iteration – the current iteration step $k$ that is is currently performed or was last performed

stopping_criterion – a StoppingCriterion that indicates whether the algorithm will stop after this iteration or has stopped.

iterate the current iterate $x^{(k)}$`.

These variable names given in this list are the defaults for which the accessors are implemented, such that if your concrete MyState <: State follows this convention, you do not have to reimplement their accessors.

Methods

The following methods should be implemented for a state

get_iteration(s) to return the current iteration number

`increment!(s)

get_stopping_criterion return the StoppingCriterion

get_iterate return the current iterate $x^{(k)}$.

AlgorithmsInterface.get_iterate — Method

get_iterate(s::State)

Return the current iterate $x^{(k)}$ of a State s

The default assumes that the current iteration is stored in s.iterate.

AlgorithmsInterface.get_iteration — Method

get_iteration(s::State)

Return the current iteration a State s either is currently performing or was last performed

The default assumes that the current iteration is stored in s.iteration.

AlgorithmsInterface.get_stopping_criterion — Method

get_stopping_criterion(s::State)

Return the StoppingCriterion of the State s.

The default assumes that the criterion is stored in s.stopping_criterion.

AlgorithmsInterface.increment! — Method

increment!(s::State)

Return the current iteration a State s either is currently performing or was last performed

The default assumes that the current iteration is stored in s.iteration.