Stopping criterion

AlgorithmsInterface.StopAfter — Type

StopAfter <: StoppingCriterion

store a threshold when to stop looking at the complete runtime. It uses time_ns() to measure the time and you provide a Period as a time limit, for example Minute(15).

Fields

threshold stores the Period after which to stop

start stores the starting time when the algorithm is started, that is a call with i=0.

time stores the elapsed time

at_iteration indicates at which iteration (including i=0) the stopping criterion was fulfilled and is -1 while it is not fulfilled.

Constructor

StopAfter(t)

initialize the stopping criterion to a Period t to stop after.

AlgorithmsInterface.StopAfterIteration — Type

StopAfterIteration <: StoppingCriterion

A functor for a stopping criterion to stop after a maximal number of iterations.

Fields

max_iterations stores the maximal iteration number where to stop at

at_iteration indicates at which iteration (including i=0) the stopping criterion was fulfilled and is -1 while it is not fulfilled.

Constructor

StopAfterIteration(maxIter)

initialize the functor to indicate to stop after maxIter iterations.

AlgorithmsInterface.StoppingCriterion — Type

StoppingCriterion

An abstract type to represent a stopping criterion of an solver.

Any concrete stopping criterion should be implemented as a functor, that takes the “usual tuple” (p, a, s) of a Problem p, an Algorithm and a State as input, where this criterion should usually be part of the State itself.

Methods

A concrete StoppingCriterion sc should provide the following functions besides the above-mentioned functor is it itself

get_reason(sc) a human readable text of about one line of length providing a reason why this stopping criterion indicated to stop. An empty string if it did not indicate to stop

get_summary(sc) a short summary of this stopping criterion, and whether it was reached, e.g. a short string like "Max Iterations (15): reached"

indicates_convergence(sc) a boolean whether or not this stopping criterion would indicate that the algorithm has converged, if it indicates to stop.

show(io::IO, sc) to display its constructor and the get_summary

AlgorithmsInterface.get_reason — Method

get_reason(sc::StoppingCriterion)

Provide a reason in human readable text as to why a StoppingCriterion indicated to stop. If it does not indicate to stop, this should return an empty string.

Providing the iteration at which this indicated to stop would be preferrable.

AlgorithmsInterface.get_summary — Method

get_summary(sc::StoppingCriterion)

Provide a summary of the status of a stopping criterion – its parameters and whether it currently indicates to stop. It should not be longer than one line

Example

For the StopAfterIteration criterion, the summary looks like

Max Iterations (15): not reached

AlgorithmsInterface.indicates_convergence — Method

indicates_convergence(sc::StoppingCriterion)

Return whether or not a StoppingCriterion indicates convergence of an algorithm if it would indicate to stop.